Fix metadata export in client pages
